Added input checking to windows

+++ MayNazMuth/AddPassengerWindow.xaml.cs
@@ -92,12 +92,21 @@ namespace MayNazMuth {
             using (var db = new CustomDbContext()) {
                 var selectedPassengers = AllPassengersDataGrid.SelectedItems;
 
-                foreach(Passenger p in selectedPassengers) {
-                    db.Passengers.Remove(p);
+                if(selectedPassengers.Count == 0) {
+                    MessageBox.Show("No Passenger selected!!");
                 }
+                else {
+                    foreach (Passenger p in selectedPassengers) {
+                        db.Passengers.Remove(p);
+                    }
+                }
+                
 
                 db.SaveChanges();
+                
                 InitializeDataGrid();
+
+                MessageBox.Show("Passenger deleted successfully");
             }
         }
